[catmem] Enhancement: Duplex Ring
